Handle strings and bool in podspecs
