+++ caas/kubernetes/provider/provider_test.go
@@ -8,9 +8,6 @@ import (
 	jc "github.com/juju/testing/checkers"
 	"github.com/juju/utils"
 	gc "gopkg.in/check.v1"
-	core "k8s.io/api/core/v1"
-	apiextensionsv1beta1 "k8s.io/apiextensions-apiserver/pkg/apis/apiextensions/v1beta1"
-	"k8s.io/apimachinery/pkg/util/intstr"
 
 	"github.com/juju/juju/caas"
 	"github.com/juju/juju/caas/kubernetes/provider"
@@ -127,236 +124,4 @@ func (s *providerSuite) TestValidate(c *gc.C) {
 
 	validAttrs := validCfg.AllAttrs()
 	c.Assert(config.AllAttrs(), gc.DeepEquals, validAttrs)
-}
-
-func (s *providerSuite) TestParsePodSpec(c *gc.C) {
-
-	specStr := `
-omitServiceFrontend: true
-containers:
-  - name: gitlab
-    image: gitlab/latest
-    imagePullPolicy: Always
-    ports:
-    - containerPort: 80
-      protocol: TCP
-    - containerPort: 443
-    livenessProbe:
-      initialDelaySeconds: 10
-      httpget:
-        path: /ping
-        port: 8080
-    readinessProbe:
-      initialDelaySeconds: 10
-      httpget:
-        path: /pingReady
-        port: www
-    config:
-      attr: foo=bar; fred=blogs
-      foo: bar
-    files:
-      - name: configuration
-        mountPath: /var/lib/foo
-        files:
-          file1: |
-            [config]
-            foo: bar
-customResourceDefinition:
-  - group: kubeflow.org
-    version: v1alpha2
-    scope: Namespaced
-    kind: TFJob
-    validation:
-      properties:
-        tfReplicaSpecs:
-          properties:
-            Worker:
-              properties:
-                replicas:
-                  type: integer
-                  minimum: 1
-            PS:
-              properties:
-                replicas:
-                  type: integer
-                  minimum: 1
-            Chief:
-              properties:
-                replicas:
-                  type: integer
-                  minimum: 1
-                  maximum: 1
-`[1:]
-
-	expectedFileContent := `
-[config]
-foo: bar
-`[1:]
-
-	k8sprovider := provider.NewProvider()
-	spec, err := k8sprovider.ParsePodSpec(specStr)
-	c.Assert(err, jc.ErrorIsNil)
-	c.Assert(spec, jc.DeepEquals, &caas.PodSpec{
-		OmitServiceFrontend: true,
-		Containers: []caas.ContainerSpec{{
-			Name:  "gitlab",
-			Image: "gitlab/latest",
-			Ports: []caas.ContainerPort{
-				{ContainerPort: 80, Protocol: "TCP"},
-				{ContainerPort: 443},
-			},
-			ProviderContainer: &provider.K8sContainerSpec{
-				ImagePullPolicy: "Always",
-				LivenessProbe: &core.Probe{
-					InitialDelaySeconds: 10,
-					Handler: core.Handler{
-						HTTPGet: &core.HTTPGetAction{
-							Path: "/ping",
-							Port: intstr.IntOrString{IntVal: 8080},
-						},
-					},
-				},
-				ReadinessProbe: &core.Probe{
-					InitialDelaySeconds: 10,
-					Handler: core.Handler{
-						HTTPGet: &core.HTTPGetAction{
-							Path: "/pingReady",
-							Port: intstr.IntOrString{StrVal: "www", Type: 1},
-						},
-					},
-				}},
-			Config: map[string]string{
-				"attr": "foo=bar; fred=blogs",
-				"foo":  "bar",
-			},
-			Files: []caas.FileSet{
-				{
-					Name:      "configuration",
-					MountPath: "/var/lib/foo",
-					Files: map[string]string{
-						"file1": expectedFileContent,
-					},
-				},
-			},
-		}},
-		CustomResourceDefinitions: []caas.CustomResourceDefinition{
-			{
-				Kind:    "TFJob",
-				Group:   "kubeflow.org",
-				Version: "v1alpha2",
-				Scope:   "Namespaced",
-				Validation: caas.CustomResourceDefinitionValidation{
-					Properties: map[string]apiextensionsv1beta1.JSONSchemaProps{
-						"tfReplicaSpecs": {
-							Properties: map[string]apiextensionsv1beta1.JSONSchemaProps{
-								"PS": {
-									Properties: map[string]apiextensionsv1beta1.JSONSchemaProps{
-										"replicas": {
-											Type: "integer", Minimum: float64Ptr(1),
-										},
-									},
-								},
-								"Chief": {
-									Properties: map[string]apiextensionsv1beta1.JSONSchemaProps{
-										"replicas": {
-											Type:    "integer",
-											Minimum: float64Ptr(1),
-											Maximum: float64Ptr(1),
-										},
-									},
-								},
-								"Worker": {
-									Properties: map[string]apiextensionsv1beta1.JSONSchemaProps{
-										"replicas": {
-											Type:    "integer",
-											Minimum: float64Ptr(1),
-										},
-									},
-								},
-							},
-						},
-					},
-				},
-			},
-		},
-	})
-}
-
-func float64Ptr(f float64) *float64 {
-	return &f
-}
-
-func (s *providerSuite) TestValidateMissingContainers(c *gc.C) {
-
-	specStr := `
-containers:
-`[1:]
-
-	_, err := provider.ParseK8sPodSpec(specStr)
-	c.Assert(err, gc.ErrorMatches, "require at least one container spec")
-}
-
-func (s *providerSuite) TestValidateMissingName(c *gc.C) {
-
-	specStr := `
-containers:
-  - image: gitlab/latest
-`[1:]
-
-	spec, err := provider.ParseK8sPodSpec(specStr)
-	c.Assert(err, jc.ErrorIsNil)
-	err = spec.Validate()
-	c.Assert(err, gc.ErrorMatches, "spec name is missing")
-}
-
-func (s *providerSuite) TestValidateMissingImage(c *gc.C) {
-
-	specStr := `
-containers:
-  - name: gitlab
-`[1:]
-
-	spec, err := provider.ParseK8sPodSpec(specStr)
-	c.Assert(err, jc.ErrorIsNil)
-	err = spec.Validate()
-	c.Assert(err, gc.ErrorMatches, "spec image details is missing")
-}
-
-func (s *providerSuite) TestValidateFileSetPath(c *gc.C) {
-
-	specStr := `
-containers:
-  - name: gitlab
-    image: gitlab/latest
-    files:
-      - files:
-          file1: |-
-            [config]
-            foo: bar
-`[1:]
-
-	spec, err := provider.ParseK8sPodSpec(specStr)
-	c.Assert(err, jc.ErrorIsNil)
-	err = spec.Validate()
-	c.Assert(err, gc.ErrorMatches, `file set name is missing`)
-}
-
-func (s *providerSuite) TestValidateMissingMountPath(c *gc.C) {
-
-	specStr := `
-containers:
-  - name: gitlab
-    image: gitlab/latest
-    files:
-      - name: configuration
-        files:
-          file1: |-
-            [config]
-            foo: bar
-`[1:]
-
-	spec, err := provider.ParseK8sPodSpec(specStr)
-	c.Assert(err, jc.ErrorIsNil)
-	err = spec.Validate()
-	c.Assert(err, gc.ErrorMatches, `mount path is missing for file set "configuration"`)
 }
