+++ worker/caasunitprovisioner/worker_test.go
@@ -77,7 +77,7 @@ containers:
 				{ContainerPort: 80, Protocol: "TCP"},
 				{ContainerPort: 443},
 			},
-			Config: map[string]string{
+			Config: map[string]interface{}{
 				"attr": "foo=bar; fred=blogs",
 				"foo":  "bar",
 			}},
